requirements.txt
numpy==1.23.5
pandas==2.0.3
matplotlib==3.7.2
seaborn==0.12.2
scikit-learn==1.3.0
scikit-surprise==1.1.3
streamlit==1.28.0
plotly==5.17.0
scipy==1.11.2
gdown==4.7.1